pyproject.toml
[tool.poetry]
name = "fox"
version = "0.1.0"
description = ""
authors = ["Ahrr Jay <rowanjohnas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.4.3"
#testproject-python-sdk = "^1.2.3"
pytest = "^7.1.2"
allure-pytest = "^2.10.0"
webdriver-manager = "^3.8.3"
assertpy = "^1.1"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
